// FrontEnd/src/app/services/api-backend/model/diagnostico.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DiagnosticosCaso } from './diagnosticosCaso';
import { DiagnosticosCitum } from './diagnosticosCitum';


export interface Diagnostico { 
    idDiagnostico?: string | null;
    nombre?: string | null;
    estado?: string | null;
    diagnosticosCasos?: Array<DiagnosticosCaso> | null;
    diagnosticosCita?: Array<DiagnosticosCitum> | null;
}