// FrontEnd/src/app/services/api-backend/model/diagnosticosCaso.ts
/**
 * Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { Diagnostico } from './diagnostico';
import { Usuario } from './usuario';
import { Caso } from './caso';


export interface DiagnosticosCaso { 
    idDiagnostiosCaso?: number;
    idCaso?: number;
    idDiagnostico?: string | null;
    observaciones?: string | null;
    fechaInicio?: string;
    fechaFin?: string | null;
    idUsuario?: number;
    estado?: string | null;
    idCasoNavigation?: Caso;
    idDiagnosticoNavigation?: Diagnostico;
    idUsuarioNavigation?: Usuario;
}